Split off phonenumbers as a plugin (WIP)
