+++ example/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls import include, url
 from django.contrib import admin
 from django.contrib.auth.views import LogoutView
 
-from two_factor.gateways.twilio.urls import urlpatterns as tf_twilio_urls
+from two_factor.plugins.phonenumber.gateways.twilio.urls import urlpatterns as tf_twilio_urls
 from two_factor.urls import urlpatterns as tf_urls
 
 from .views import (
