+++ two_factor/admin.py
@@ -6,7 +6,6 @@ from django.contrib.auth.views import redirect_to_login
 from django.shortcuts import resolve_url
 from django.utils.http import is_safe_url
 
-from .models import PhoneDevice
 from .utils import monkeypatch_method
 
 
@@ -65,14 +64,3 @@ def unpatch_admin():
 
 
 original_login = AdminSite.login
-
-
-class PhoneDeviceAdmin(admin.ModelAdmin):
-    """
-    :class:`~django.contrib.admin.ModelAdmin` for
-    :class:`~two_factor.models.PhoneDevice`.
-    """
-    raw_id_fields = ('user',)
-
-
-admin.site.register(PhoneDevice, PhoneDeviceAdmin)
