+++ two_factor/forms.py
@@ -8,10 +8,11 @@ from django_otp.forms import OTPAuthenticationFormMixin
 from django_otp.oath import totp
 from django_otp.plugins.otp_totp.models import TOTPDevice
 
-from .models import (
-    PhoneDevice, get_available_methods, get_available_phone_methods,
-)
-from .utils import totp_digits
+#from .models import (
+    #PhoneDevice,
+    #get_available_phone_methods,
+#)
+from .utils import totp_digits, get_available_methods
 from .validators import validate_international_phonenumber
 
 try:
@@ -28,30 +29,6 @@ class MethodForm(forms.Form):
     def __init__(self, **kwargs):
         super(MethodForm, self).__init__(**kwargs)
         self.fields['method'].choices = get_available_methods()
-
-
-class PhoneNumberMethodForm(ModelForm):
-    number = forms.CharField(label=_("Phone Number"),
-                             validators=[validate_international_phonenumber])
-    method = forms.ChoiceField(widget=forms.RadioSelect, label=_('Method'))
-
-    class Meta:
-        model = PhoneDevice
-        fields = 'number', 'method',
-
-    def __init__(self, **kwargs):
-        super(PhoneNumberMethodForm, self).__init__(**kwargs)
-        self.fields['method'].choices = get_available_phone_methods()
-
-
-class PhoneNumberForm(ModelForm):
-    # Cannot use PhoneNumberField, as it produces a PhoneNumber object, which cannot be serialized.
-    number = forms.CharField(label=_("Phone Number"),
-                             validators=[validate_international_phonenumber])
-
-    class Meta:
-        model = PhoneDevice
-        fields = 'number',
 
 
 class DeviceValidationForm(forms.Form):
@@ -139,11 +116,16 @@ class DisableForm(forms.Form):
 
 
 class AuthenticationTokenForm(OTPAuthenticationFormMixin, Form):
-    otp_token = forms.IntegerField(label=_("Token"), min_value=1,
-                                   max_value=int('9' * totp_digits()))
-
-    otp_token.widget.attrs.update({'autofocus': 'autofocus'})
-
+    # otp_token = forms.IntegerField(label=_("Token"), min_value=1,
+    #                                max_value=int('9' * totp_digits()))
+    otp_token = forms.RegexField(label=_("Token"),
+                                 regex=r'^[0-9]*$',
+                                 min_length=totp_digits(),
+                                 max_length=totp_digits())
+    otp_token.widget.attrs.update({
+        'autofocus': 'autofocus',
+        'pattern': '[0-9]*', # hint to show numeric keyboard for on-screen keyboards
+    })
     # Our authentication form has an additional submit button to go to the
     # backup token form. When the `required` attribute is set on an input
     # field, that button cannot be used on browsers that implement html5
