+++ two_factor/views/profile.py
@@ -6,8 +6,8 @@ from django.views.generic import FormView, TemplateView
 from django_otp import devices_for_user, user_has_device
 
 from ..forms import DisableForm
-from ..models import get_available_phone_methods
-from ..utils import backup_phones, default_device
+# from ..models import get_available_phone_methods
+# from ..utils import backup_phones, default_device
 from .utils import class_view_decorator
